icard_django/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
idna==3.6
inflection==0.5.1
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
